+++ Tests/FeatherDatabaseMySQLTests/FeatherDatabaseMySQLTestSuite.swift
@@ -1,6 +1,6 @@
 //
-//  FeatherMySQLDatabaseTestSuite.swift
-//  feather-mysql-database
+//  FeatherDatabaseMySQLTestSuite.swift
+//  feather-database-mysql
 //
 //  Created by Tibor Bödecs on 2026. 01. 10..
 //
@@ -13,7 +13,7 @@ import NIOPosix
 import NIOSSL
 import Testing
 
-@testable import FeatherMySQLDatabase
+@testable import FeatherDatabaseMySQL
 
 #if canImport(FoundationEssentials)
 import FoundationEssentials
@@ -22,7 +22,7 @@ import Foundation
 #endif
 
 @Suite
-struct MySQLDatabaseTestSuite {
+struct FeatherDatabaseMySQLTestSuite {
 
     func randomTableSuffix() -> String {
         let characters = Array("abcdefghijklmnopqrstuvwxyz0123456789")
@@ -35,7 +35,7 @@ struct MySQLDatabaseTestSuite {
     }
 
     func runUsingTestDatabaseClient(
-        _ closure: ((MySQLDatabaseClient) async throws -> Void)
+        _ closure: ((DatabaseClientMySQL) async throws -> Void)
     ) async throws {
         var logger = Logger(label: "test")
         logger.logLevel = .info
@@ -69,7 +69,7 @@ struct MySQLDatabaseTestSuite {
             )
             .get()
 
-        let database = MySQLDatabaseClient(
+        let database = DatabaseClientMySQL(
             connection: connection,
             logger: logger
         )
@@ -1561,6 +1561,170 @@ struct MySQLDatabaseTestSuite {
                     try result[0].decode(column: "is_valid", as: Bool.self)
                         == true
                 )
+            }
+        }
+    }
+
+    // MARK: - sequence tests
+
+    @Test
+    func rowSequenceIteratesRowsInOrder() async throws {
+        try await runUsingTestDatabaseClient { database in
+            let suffix = randomTableSuffix()
+            let table = "planets_\(suffix)"
+
+            try await database.withConnection { connection in
+                try await connection.run(
+                    query: #"""
+                        DROP TABLE IF EXISTS `\#(unescaped: table)`;
+                        """#
+                )
+                try await connection.run(
+                    query: #"""
+                        CREATE TABLE IF NOT EXISTS `\#(unescaped: table)` (
+                            `id` INTEGER PRIMARY KEY,
+                            `name` TEXT
+                        );
+                        """#
+                )
+
+                try await connection.run(
+                    query: #"""
+                        INSERT INTO `\#(unescaped: table)` (`id`, `name`)
+                        VALUES
+                            (1, 'Mercury'),
+                            (2, 'Venus');
+                        """#
+                )
+
+                let sequence = try await connection.run(
+                    query: #"""
+                        SELECT *
+                        FROM `\#(unescaped: table)`
+                        ORDER BY `id` ASC;
+                        """#
+                )
+
+                var iterator = sequence.makeAsyncIterator()
+
+                let first = await iterator.next()
+                #expect(first != nil)
+                #expect(
+                    try first?.decode(column: "name", as: String.self)
+                        == "Mercury"
+                )
+
+                let second = await iterator.next()
+                #expect(second != nil)
+                #expect(
+                    try second?.decode(column: "name", as: String.self)
+                        == "Venus"
+                )
+
+                let third = await iterator.next()
+                #expect(third == nil)
+            }
+        }
+    }
+
+    @Test
+    func rowSequenceCollectReturnsAllRows() async throws {
+        try await runUsingTestDatabaseClient { database in
+            let suffix = randomTableSuffix()
+            let table = "greetings_\(suffix)"
+
+            try await database.withConnection { connection in
+                try await connection.run(
+                    query: #"""
+                        DROP TABLE IF EXISTS `\#(unescaped: table)`;
+                        """#
+                )
+                try await connection.run(
+                    query: #"""
+                        CREATE TABLE IF NOT EXISTS `\#(unescaped: table)` (
+                            `id` INTEGER PRIMARY KEY,
+                            `name` TEXT
+                        );
+                        """#
+                )
+
+                try await connection.run(
+                    query: #"""
+                        INSERT INTO `\#(unescaped: table)` (`id`, `name`)
+                        VALUES
+                            (1, 'Hello'),
+                            (2, 'World');
+                        """#
+                )
+
+                let sequence = try await connection.run(
+                    query: #"""
+                        SELECT
+                            `id`,
+                            `name`
+                        FROM `\#(unescaped: table)`
+                        ORDER BY `id` ASC;
+                        """#
+                )
+
+                let rows = try await sequence.collect()
+                #expect(rows.count == 2)
+
+                let firstName = try rows[0]
+                    .decode(
+                        column: "name",
+                        as: String.self
+                    )
+                let secondName = try rows[1]
+                    .decode(
+                        column: "name",
+                        as: String.self
+                    )
+
+                #expect(firstName == "Hello")
+                #expect(secondName == "World")
+            }
+        }
+    }
+
+    @Test
+    func rowSequenceHandlesEmptyResults() async throws {
+        try await runUsingTestDatabaseClient { database in
+            let suffix = randomTableSuffix()
+            let table = "empty_rows_\(suffix)"
+
+            try await database.withConnection { connection in
+                try await connection.run(
+                    query: #"""
+                        DROP TABLE IF EXISTS `\#(unescaped: table)`;
+                        """#
+                )
+                try await connection.run(
+                    query: #"""
+                        CREATE TABLE IF NOT EXISTS `\#(unescaped: table)` (
+                            `id` INTEGER PRIMARY KEY,
+                            `name` TEXT
+                        );
+                        """#
+                )
+
+                let sequence = try await connection.run(
+                    query: #"""
+                        SELECT
+                            `id`,
+                            `name`
+                        FROM `\#(unescaped: table)`
+                        WHERE
+                            1=0;
+                        """#
+                )
+
+                let rows = try await sequence.collect()
+                #expect(rows.isEmpty)
+
+                var iterator = sequence.makeAsyncIterator()
+                let first = await iterator.next()
+                #expect(first == nil)
             }
         }
     }
